[FL-3677, FL-3798] RFID Improvements (#3524)
* Update saved_info and read_success scenes
* Update EM4100 rendering
* Update HIDExt rendering
* Update Gallagher rendering
* Update HidProx rendering
* Update IOProx rendering
* Update H10301 rendering
* Update PAC/Stanley rendering
* Add strcasecmp() to API, better manufacturer/name handling
* Update Viking rendering
* Update FDX-A rendering
* Update Pyramid rendering
* Update Indala26 rendering
* Update Idteck rendering
* Update Keri rendering
* Update Nexwatch rendering
* Update Jablotron rendering
* Update Paradox rendering
* Truncate long Hex string on scene_read_suceess
* Fix formatting
* Update AWID rendering
* Update FDX-B rendering
* Tweak string formatting in various screens
* More read_success view tweaks
* Fix formatting
* Fix Pyramid brief rendering
* Reset saved key menu when going back
* Reset other menus on back where applicable
* Update confirmation scenes
* Update emulation scene
* Update delete scene
* Update raw read info screen
* Update raw read scene, fix crash
* Update raw read success scene
* Update write scene
* Always return to SceneSelectKey after saving
* Update SceneWriteSuccess and SceneDeleteSuccess
* Replace closing parens with dots
* FL-3798: Fix special formatting in text_box
* Simplify SceneReadSuccess
* Fix crash when having a trailing newline in text_box
* Bump API symbols version
* Make PVS happy
* Format sources

+++ applications/main/nfc/helpers/protocol_support/st25tb/st25tb_render.c
@@ -13,15 +13,16 @@ void nfc_render_st25tb_info(
     }
 
     if(format_type == NfcProtocolFormatTypeFull) {
-        furi_string_cat_printf(str, "\nSys. OTP: %08lX", __bswap32(data->system_otp_block));
+        furi_string_cat_printf(
+            str, "\nSys. OTP: %08lX", (uint32_t)__bswap32(data->system_otp_block));
         furi_string_cat_printf(str, "\n::::::::::::::::::::::[Blocks]::::::::::::::::::::::");
         for(size_t i = 0; i < st25tb_get_block_count(data->type); i += 2) {
             furi_string_cat_printf(
                 str,
                 "\n %02X   %08lX  %08lX",
                 i,
-                __bswap32(data->blocks[i]),
-                __bswap32(data->blocks[i + 1]));
+                (uint32_t)__bswap32(data->blocks[i]),
+                (uint32_t)__bswap32(data->blocks[i + 1]));
         }
     }
 }
